feat: update battlescribe-spec submodule and implement collective semantics
Update battlescribe-spec submodule to latest main (4e93ef3).

Conformance changes (351/353 passing, up from 322/353):
- Fix double->decimal migration for TestKit Protocol types
- Implement entry-id prefix propagation for link chains
- Fix roster name from test context
- Implement multiple additive repeats
- Implement includeChildForces scope traversal
- Implement SEG constraint evaluation
- Implement shared flag for constraints/conditions/repeats
- Fix selection ordering (effective name sort with natural ordering)
- Fix PrimaryCatalogue scope in BindFilterEntrySymbol

Collective feature (all 18 collective specs passing):
- Surface IsCollective on ISelectionEntryContainerSymbol interface
- SetSelectionCountById: per-model semantics for collective entries,
  recursive child number propagation on parent count changes
- SelectChildEntryById: collective children inherit parent number,
  isDuplicate mode for entries with only collective/hidden children
- DeselectSelectionById: per-model deselect for collective entries
- ConstraintEvaluator: per-model constraint counting for collective

Remaining 2 failures:
- scope-child-id-filter: genuine ordering conflict (needs spec override)
- protocol-kitchen-sink: hidden diagnostic + force count (pre-existing)

+++ src/WarHub.ArmouryModel.RosterEngine.Spec/ProtocolConverter.cs
@@ -215,7 +215,7 @@ public static class ProtocolConverter
     {
         Id = c.Id,
         Type = ParseConstraintKind(c.Type),
-        Value = (decimal)c.Value,
+        Value = c.Value,
         Field = c.Field,
         Scope = c.Scope,
         Shared = c.Shared,
@@ -246,7 +246,7 @@ public static class ProtocolConverter
     private static ConditionCore ConvertCondition(ProtocolCondition c) => new()
     {
         Type = ParseConditionKind(c.Type),
-        Value = (decimal)c.Value,
+        Value = c.Value,
         Field = c.Field,
         Scope = c.Scope,
         ChildId = c.ChildId,
@@ -265,7 +265,7 @@ public static class ProtocolConverter
 
     private static RepeatCore ConvertRepeat(ProtocolRepeat r) => new()
     {
-        Value = (decimal)r.Value,
+        Value = r.Value,
         RepeatCount = r.Repeats,
         Field = r.Field,
         Scope = r.Scope,
@@ -355,7 +355,7 @@ public static class ProtocolConverter
     {
         Id = ct.Id,
         Name = ct.Name,
-        DefaultCostLimit = ct.DefaultCostLimit is { } limit ? (decimal)limit : -1m,
+        DefaultCostLimit = ct.DefaultCostLimit is { } limit ? limit : -1m,
         Hidden = ct.Hidden,
     };
 
@@ -363,7 +363,7 @@ public static class ProtocolConverter
     {
         Name = cv.Name,
         TypeId = cv.TypeId,
-        Value = (decimal)cv.Value,
+        Value = cv.Value,
     };
 
     private static ProfileTypeCore ConvertProfileType(ProtocolProfileType pt) => new()
